prepare for git-attribute file parsing (#301)
