+++ git-attributes/src/parse/mod.rs
@@ -1,5 +1,7 @@
 pub mod ignore;
 
+pub mod attributes {}
+
 pub fn ignore(buf: &[u8]) -> ignore::Iter<'_> {
     ignore::Iter::new(buf)
 }
