+++ git-attributes/tests/parse/mod.rs
@@ -1,0 +1,1 @@
+mod ignore;
